from: https://sergiistarodubtsev.substack.com/
"A story about a bear's perspective on what it means to be human (short)"

About biases:

Our minds have a bunch of tricks and biases. For example, we often focus on recent events and forget older ones. Or we might believe something is true just because lots of people say so. These biases make us see things in a certain way, even if it's not really true.

All these thinkings kinda make it clear that reality isn't all cut and dried, you know. Our minds, the things we believe, and those sneaky little things called biases, they all shape the way we look at the world. It's a bit like we're all snugly pieces in a grand puzzle, all tryin' to work out what's real and what's not. It's like we're off on a curious adventure, full of “A-ha!” moments.....